+++ app/assets/stylesheets/colors.scss
@@ -1,2 +1,3 @@
 $blue-00aeef: #00aeef;
 $gray-ddd: #ddd;
+$gray-e0: #e0e0e075;
