+++ app/javascript/packs/application.js
@@ -14,7 +14,7 @@ require('cocoon')
 require('packs/custom')
 require('chartkick')
 require('chart.js')
-// require('packs/viewed_noti.js')
+require('packs/count_association')
 
 import '@fortawesome/fontawesome-free/js/all';
 import $ from 'jquery';
