+++ app/javascript/packs/count_association.js
@@ -1,0 +1,9 @@
+$(function(){
+  $('.add_fields').click(function(){
+    let field = $(this).data('associations');
+    setTimeout(() => {
+      let newBoxItem = $(`.${field} .box-item`).last()
+      newBoxItem[0].querySelector('.total_item').innerHTML = newBoxItem.index() + 1
+    }, 0)
+  })
+})
