+++ app/views/layouts/application.html.erb
@@ -15,7 +15,7 @@
 
   <body>
     <%= render "layouts/header" %>
-    <%= render "shared/notice_messages" %>
+    <%= custom_flash %>
     <%= yield %>
     <%= render "layouts/footer" %>
