+++ app/views/profiles/_experience_fields.html.erb
@@ -1,12 +1,25 @@
-<div class="nested-fields">
-  <div class="resumeInfo">
-    <div class="head">
+<div class="box-item nested-fields">
+  <div class="row">
+    <div class="col col-sm-12 col-md-12 col-lg-12">
+      <div class="close-nested-field"> 
+        <span class="total_item"><%= Settings.index_number %></span>
+        <%= link_to_remove_association f, class: "btn-sm btn-danger" do %>
+          <i class="fas fa-times"></i>
+        <% end %>
+      </div>
+    </div>
+  </div>
+  <div class="row">
+    <div class="col col-sm-12 col-md-12 col-lg-12">
+      <%= f.label :title, t("cv.title") %>
       <%= f.text_field :title, placeholder: t("cv.title"), class: "form-control" %>
       <%= errors_for f, :title %>
     </div>
   </div>
-  <%= f.text_area :description, class: "form-control" %>
-  <%= link_to_remove_association f do %>
-    <i class="fas fa-minus-circle text-danger"></i>
-  <% end %>
+  <div class="row mt-10">
+    <div class="col col-sm-12 col-md-12 col-lg-12">
+      <%= f.label :description, t("cv.content") %>
+      <%= f.text_area :description, class: "form-control" %>
+    </div>
+  </div>
 </div>
